+++ modules/special_functions/builtin/cpp/erfcxBuiltin.cpp
@@ -1,0 +1,36 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "erfcxBuiltin.hpp"
+#include "Error.hpp"
+#include "ErrorFunctions.hpp"
+#include "OverloadRequired.hpp"
+#include "i18n.hpp"
+#include "InputOutputArgumentsCheckers.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::SpecialFunctionsGateway::erfcxBuiltin(int nLhs, const ArrayOfVector& argIn)
+{
+    ArrayOfVector retval;
+    nargincheck(argIn, 1, 1);
+    nargoutcheck(nLhs, 0, 1);
+    if (argIn[0].isSparse() || argIn[0].getDataClass() == NLS_DCOMPLEX
+        || argIn[0].getDataClass() == NLS_SCOMPLEX) {
+        Error(_W("Input argument must be dense and real."));
+    }
+    if (argIn[0].getDataClass() == NLS_DOUBLE || argIn[0].getDataClass() == NLS_SINGLE) {
+        retval << ErrorFunction(argIn[0], ErrorFunctionType::Erfcx);
+    } else {
+        OverloadRequired("erfcx");
+    }
+    return retval;
+}
+//=============================================================================